+++ examples/sumomo/src/sumomo.cpp
@@ -37,6 +37,7 @@ struct MomoSampleConfig {
   boost::optional<bool> spotlight;
   int spotlight_number = 0;
   boost::optional<bool> simulcast;
+  boost::optional<bool> simulcast_multicodec;
   boost::optional<bool> data_channel_signaling;
   boost::optional<bool> ignore_disconnect_websocket;
 
@@ -142,6 +143,7 @@ class MomoSample : public std::enable_shared_from_this<MomoSample>,
     config.spotlight = config_.spotlight;
     config.spotlight_number = config_.spotlight_number;
     config.simulcast = config_.simulcast;
+    config.simulcast_multicodec = config_.simulcast_multicodec;
     config.data_channel_signaling = config_.data_channel_signaling;
     config.ignore_disconnect_websocket = config_.ignore_disconnect_websocket;
     config.proxy_agent = "Momo Sample for Sora C++ SDK";
@@ -375,6 +377,8 @@ int main(int argc, char* argv[]) {
       ->check(CLI::Range(0, 8));
   add_optional_bool(app, "--simulcast", config.simulcast,
                     "Use simulcast (default: none)");
+  add_optional_bool(app, "--simulcast-multicodec", config.simulcast_multicodec,
+                    "Use simulcast multi-codec (default: none)");
   add_optional_bool(app, "--data-channel-signaling",
                     config.data_channel_signaling,
                     "Use DataChannel for Sora signaling (default: none)");